pyproject.toml
[project]
name = "ralph-wiggum-experiment-tools"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pyyaml>=6.0",
]
